ZF2\EventManager: Corrections to catching exceptions.
Add code showing that it is not an error to trigger an event that has not been registered with the EventManager.
Add exit() to terminate the script if an exception is thrown.
Catch Zend\Stdlib\Exception\InvalidCallbackException if attach is given a bad callback. attach() should probably call Zend\EventManager\Exception\InvalidCallbackException, but does not.